separation of calc.cjs into util methods and tests written for those utils
